cdf-monitoring/requirements.txt
requests
google.auth
google-cloud-storage
google-cloud-bigquery